Corrected typing mistake XD

+++ Quakaster/core/utility/Service.h
@@ -5,12 +5,12 @@ namespace qk {
 	struct QK_API IService {
 
 		protected:
-			size_t m_Uptime_ns = 0;
+			uint64_t m_Uptime_ns = 0;
 
 
 		public:
-			size_t* uptime_ptr();
-			size_t uptime_ns();
+			uint64_t* uptime_ptr();
+			uint64_t uptime_ns();
 
 			virtual ~IService() = default;
 			virtual void init();
